chatbot-project/chatbot/pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
langchain-openai = "^0.1.8"
langchain-community = "^0.2.1"
langchain-core = "^0.2.3"
langchain-text-splitters = "^0.2.0"
langchain-chroma = "^0.1.1"
unstructured = {extras = ["local-inference"], version = "^0.14.3"}
langchainhub = "^0.1.17"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
